rebasing branch to main

+++ entities/src/main/java/org/entities/Cart.java
@@ -6,16 +6,17 @@ import java.util.List;
 
 
 public class Cart {
-
     private List<Product> products = new LinkedList<>();
-    private double totalCost;
+    private int totalCost;
 
     public Cart() {
         this.products = new LinkedList<Product>();
         this.totalCost = 0;
     }
 
-    public double getTotalCost() {
+
+
+    public int getTotalCost() {
         return this.totalCost;
     }
 
@@ -23,7 +24,7 @@ public class Cart {
         return this.products;
     }
 
-    public void setTotalCost(double newCost) {
+    public void setTotalCost(int newCost) {
         this.totalCost = newCost;
     }
